requirements.txt
asarPy==1.0.1
bcrypt==4.2.1
bidict==0.23.1
blinker==1.9.0
click==8.1.8
dnspython==1.16.0
eventlet==0.33.3
Flask==3.1.0
Flask-Bcrypt==1.0.1
Flask-Cors==5.0.0
Flask-SocketIO==5.3.3
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
gunicorn==23.0.0
h11==0.14.0
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
packaging==24.2
pymongo==3.12.0
python-dotenv==1.0.1
python-engineio==4.11.2
python-socketio==5.12.1
simple-websocket==1.1.0
six==1.17.0
SQLAlchemy==2.0.36
typing_extensions==4.12.2
Werkzeug==3.1.3
wsproto==1.2.0
